FRI and basic FRI separated. #20
