+++ test/commitment/fri.cpp
@@ -74,7 +74,7 @@ BOOST_AUTO_TEST_CASE(fri_basic_test) {
     constexpr static const std::size_t d_extended = d;
     std::size_t extended_log = boost::static_log2<d_extended>::value;
     std::vector<std::shared_ptr<math::evaluation_domain<FieldType>>> D =
-        fri_type::calculate_domain_set(extended_log, r);
+        zk::commitments::detail::calculate_domain_set<FieldType>(extended_log, r);
 
     math::polynomial<typename FieldType::value_type> q = {0, 0, 1};
     params.r = r;
@@ -96,17 +96,12 @@ BOOST_AUTO_TEST_CASE(fri_basic_test) {
     std::vector<std::uint8_t> init_blob {0, 1, 2, 3, 4, 5, 6, 7, 8, 9};
     zk::transcript::fiat_shamir_heuristic_sequential<transcript_hash_type> transcript(init_blob);
 
-    // LPC-related logic, here we "nulify" it via U = 0, V - 1
-    // TODO: Make FRI independent from LPC input
-    math::polynomial<typename FieldType::value_type> U = {0};
-    math::polynomial<typename FieldType::value_type> V = {1};
-
-    proof_type proof = fri_type::proof_eval(f, f, commit_merkle, params, transcript);
+    proof_type proof = fri_type::proof_eval(f, commit_merkle, params, transcript);
 
     // verify
     zk::transcript::fiat_shamir_heuristic_sequential<transcript_hash_type> transcript_verifier(init_blob);
 
-    BOOST_CHECK(fri_type::verify_eval(proof, params, U, V, transcript_verifier));
+    BOOST_CHECK(fri_type::verify_eval(proof, params, transcript_verifier));
 
     typename FieldType::value_type verifier_next_challenge = transcript_verifier.template challenge<FieldType>();
     typename FieldType::value_type prover_next_challenge = transcript.template challenge<FieldType>();
@@ -136,7 +131,8 @@ BOOST_AUTO_TEST_CASE(fri_fold_test) {
     math::polynomial<typename FieldType::value_type> q = {0, 0, 1};
 
     std::size_t d_log = boost::static_log2<d>::value;
-    std::vector<std::shared_ptr<math::evaluation_domain<FieldType>>> D = fri_type::calculate_domain_set(d_log, 1);
+    std::vector<std::shared_ptr<math::evaluation_domain<FieldType>>> D = 
+        zk::commitments::detail::calculate_domain_set<FieldType>(d_log, 1);
 
     params.r = r;
     params.D = D;
@@ -149,7 +145,7 @@ BOOST_AUTO_TEST_CASE(fri_fold_test) {
     typename FieldType::value_type x_next = params.q.evaluate(omega);
     typename FieldType::value_type alpha = algebra::random_element<FieldType>();
 
-    math::polynomial<typename FieldType::value_type> f_next = fri_type::fold_polynomial(f, alpha);
+    math::polynomial<typename FieldType::value_type> f_next = zk::commitments::detail::fold_polynomial<FieldType>(f, alpha);
 
     BOOST_CHECK_EQUAL(f_next.degree(), f.degree() / 2);
     std::vector<std::pair<typename FieldType::value_type, typename FieldType::value_type>> interpolation_points {
@@ -189,7 +185,7 @@ BOOST_AUTO_TEST_CASE(fri_steps_count_test) {
     constexpr static const std::size_t d_extended = d;
     std::size_t extended_log = boost::static_log2<d_extended>::value;
     std::vector<std::shared_ptr<math::evaluation_domain<FieldType>>> D =
-        fri_type::calculate_domain_set(extended_log, r);
+        zk::commitments::detail::calculate_domain_set<FieldType>(extended_log, r);
 
     params.r = r - 1;
     params.D = D;
@@ -203,7 +199,7 @@ BOOST_AUTO_TEST_CASE(fri_steps_count_test) {
     std::vector<std::uint8_t> init_blob {0, 1, 2, 3, 4, 5, 6, 7, 8, 9};
     zk::transcript::fiat_shamir_heuristic_sequential<transcript_hash_type> transcript(init_blob);
 
-    proof_type proof = fri_type::proof_eval(f, f, commit_merkle, params, transcript);
+    proof_type proof = fri_type::proof_eval(f, commit_merkle, params, transcript);
 
     math::polynomial<typename FieldType::value_type> final_polynomial = proof.final_polynomial;
     BOOST_CHECK_EQUAL(proof.final_polynomial.degree(), 1);
