+++ test/commitment/lpc.cpp
@@ -79,7 +79,7 @@ BOOST_AUTO_TEST_CASE(lpc_basic_test) {
     constexpr static const std::size_t d_extended = d;
     std::size_t extended_log = boost::static_log2<d_extended>::value;
     std::vector<std::shared_ptr<math::evaluation_domain<FieldType>>> D =
-        fri_type::calculate_domain_set(extended_log, r);
+        zk::commitments::detail::calculate_domain_set<FieldType>(extended_log, r);
 
     typename fri_type::params_type fri_params;
